requirement.txt
bidict==0.23.1
blinker==1.8.2
cffi==1.17.0
click==8.1.7
colorama==0.4.6
connect4==0.0.3
Flask==3.0.3
Flask-Cors==4.0.1
Flask-JWT-Extended==4.6.0
Flask-SocketIO==5.3.6
gevent==24.2.1
greenlet==3.0.3
h11==0.14.0
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==2.1.5
psycopg2==2.9.9
pycparser==2.22
PyJWT==2.9.0
python-engineio==4.9.1
python-socketio==5.11.3
PyYAML==6.0.2
setuptools==72.2.0
simple-websocket==1.0.0
watchdog==4.0.2
websocket==0.2.1
websockets==12.0
Werkzeug==3.0.3
wsproto==1.2.0
zope.event==5.0
zope.interface==7.0.1
